Refactor the autosuggest listeners so that it matches the rest of the save and restore input helpers
